exp4j library update

+++ app/src/main/java/org/solovyev/android/plotter/math/ExpressionFunction.java
@@ -51,14 +51,14 @@ public final class ExpressionFunction extends Function {
 
     @Override
     public float evaluate(float x) {
-        expression.setVariable(variables[0], (double) x);
+        expression.setVariable(variables[0], x);
         return (float) expression.evaluate();
     }
 
     @Override
     public float evaluate(float x, float y) {
-        expression.setVariable(variables[0], (double) x);
-        expression.setVariable(variables[1], (double) y);
+        expression.setVariable(variables[0], x);
+        expression.setVariable(variables[1], y);
         return (float) expression.evaluate();
     }
